Create README.md
This app allows the users to select a dog breed and then view one, three or five random photo's of that breed at a time. Users can also hit a "random" button that calls a single dog photo from the Dog API to the front-end. I used Python, Flask, and HTTP methods. I converted a JSON data file from an API into a dictionary that is compatible with the app. Also used Jinja placeholders to send data users request to the HTML template.